refac: login 버튼 재사용
